refactor(cargo-shuttle): break suggestions module
