+++ cargo-shuttle/src/suggestions/mod.rs
@@ -1,0 +1,6 @@
+pub mod api_key;
+pub mod deploy;
+pub mod deployment;
+pub mod logs;
+pub mod project;
+pub mod resources;
